+++ src/App.jsx
@@ -1,17 +1,21 @@
 import Hero from "./components/hero/hero";
 import ParticlesComponent from "./components/heroConfig/particles_Config";
+import NavBar from "./components/navBar";
+
 import Projects from "./components/projects";
+
+import cardSection from "./components/cardSection";
 import CardSection from "./components/cardSection";
-import Navbar from "./components/navbar/navbar";
+
 
 function App() {
   return (
     <>
-      <Navbar />
+      <NavBar />
       <ParticlesComponent id="particles" />
-      <Hero />
-      <CardSection />
+      <CardSection/>
       <Projects />
+
     </>
   );
 }
